jb-lib/pyproject.toml
[tool.poetry]
name = "lib"
version = "0.1.0"
description = ""
authors = ["Shrey Pandey <shreypandey1509@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
confluent-kafka = "^2.3.0"
azure-storage-blob = "^12.19.0"
aiofiles = "^23.1.0"
asyncpg = "^0.29.0"
pytz = "^2023.3.post1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.10.0"
google-cloud-storage = "^2.18.2"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
pytest-mock = "*"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./"